pyproject.toml
[tool.poetry]
name = "arxiv-nlu-playground"
version = "0.1.0"
description = "Developing NLU models that can interpret the contents of research papers."
authors = ["Armand Silviu Gurgu <armandgurgu23@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
numpy = "^1.21.3"
pyspark = "^3.2.0"
pytest = "^6.2.5"
textdistance = "^4.2.1"
hydra-core = "^1.1.1"
ipython = "^7.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
